Changes in the Contact section / navbar
fixed the link so it actually uses the correct email. Changed the mail
icon from bounce to pulse, less annoying.

Substitute for the Raleway font, cleaner and skinnier look (especially
since there is a lot of text in the service sections)

removed the border line from the NavBar and added breaks/space between
buttons in each section and the photo below.